-Altered Type and Reference Controllers so that they do not allow deleting when their are instances of CIs/Statuses/Relations. -Added Validation Messages for the above -Altered Domain Objects so that cascading deletes does not happen for Types -Tested All of the above -Altered Project Controller to not allow delete when there are environments -Added validation message to the above -Tested the above -Altered show Status Reference page to show configuration Items used by references -Fixed broken Results link for Test Results

+++ trunk/CDR/grails-app/controllers/ProjectController.groovy
@@ -1,4 +1,4 @@
-class ProjectController{
+class ProjectController {
     /*def scaffold = Project*/
     static accessControl = {
         // All actions require the 'Observer' role.
@@ -11,19 +11,24 @@ class ProjectController{
 
     def list = {
         if (!params.max) params.max = 10
-                [projectList: Project.list(params)]
+        [projectList: Project.list(params)]
     }
 
     def show = {
-                [project: Project.get(params.id)]
+        [project: Project.get(params.id)]
     }
 
     def delete = {
         def project = Project.get(params.id)
         if (project) {
-            project.delete()
-            flash.message = "Project ${params.id} deleted."
-            redirect(action: list)
+            if (!project.environments) {
+                project.delete()
+                flash.message = "Project ${params.id} deleted."
+                redirect(action: list)
+            } else {
+                flash.message = "Unable to Delete Project ${project.name}.  Projects with environments cannot be deleted."
+                redirect(action:show, id:params.id)
+            }
         }
         else {
             flash.message = "Project not found with id ${params.id}"
